requirements.txt
PyYAML
Telethon